- split SPK into logic and renderer
